py_version/pyproject.toml
[project]
name = "temporal-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Vikram Khandelwal",email = "vikram@runtrellis.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "temporalio (==1.17.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "openai (>=1.106.1,<2.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
temporalio = "1.17.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
